requirements.txt
hydra-core>=1.3.0
hydra-colorlog>=1.2.0
omegaconf>=2.3.0
pytorch3d>=0.7.0
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
tqdm>=4.65.0
transformers>=4.28.0
open3d>=0.17.0
scikit-learn>=1.2.0
dataclasses-json>=0.5.7
pandas>=2.0.0
trimesh>=3.15.0
structlog>=23.1.0
pytransform3d>=1.15.0
py-rouge>=1.1
wandb>=0.15.0
tenacity>=8.2.0 